pyproject.toml
[tool.poetry]
name = "pois-processor"
version = "1.0.0"
description = "Processing engine to map geohashes to POIs"
authors = ["Ignacio"]
readme = "README.md"
repository = "https://github.com/ignacio-ibanez/pois-processor"

[tool.poetry.dependencies]
python = "^3.8.12"
poetry = "^1.1.11"
pyspark = "3.2.1"
python-geohash = "0.8.5"
Shapely = "1.8.0"
geolib = "1.0.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.10b0"
pre-commit = "^2.15.0"
isort = "^5.10.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
